requirements.txt
absl-py==2.1.0
accelerate==0.26.1
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
clip-interrogator==0.6.0
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.0
decorator==5.1.1
diffusers==0.25.1
distlib==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
ExifRead-nocycle==3.0.1
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fire==0.5.0
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
huggingface-hub==0.20.2
identify==2.5.33
idna==3.6
imageio==2.33.1
img2dataset==1.45.0
imgaug==0.4.0
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
lmdb==1.4.1
lpips==0.1.4
lz4==4.3.3
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.8.0
numba==0.58.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
openai-clip==1.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyiqa==0.1.10
scikit_learn==1.7.1
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
sentencepiece==0.1.99
sentry-sdk==1.42.0
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchmetrics==1.3.1
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
virtualenv==20.25.0
wandb==0.16.4
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0